refactor: Use `u16` for `start_idx` in `subscribe_to_next_responses`

+++ crates/concepts/src/storage.rs
@@ -979,7 +979,7 @@ pub trait DbConnection: DbExecutor {
     async fn subscribe_to_next_responses(
         &self,
         execution_id: &ExecutionId,
-        start_idx: usize,
+        start_idx: u16,
         timeout_fut: Pin<Box<dyn Future<Output = ()> + Send>>,
     ) -> Result<Vec<JoinSetResponseEventOuter>, DbErrorReadWithTimeout>;
